Create server.php
The anagram server is a system that responds to subject-based search of anagrams over internet. Basically, once the keyed subject is provided, the server starts a matching process and therefore, retrieve the matches (if any). The matching process can be anything of any technical origin. The most common one, which is the one we'll be mounting later, is a search within a dictionary of anagrams. In this scenario, the patterns (the composing characters) of the subject are stored in a regular expression that is later used to do the anagram matching with the words/phrases/expressions in the dictionary.